services/device-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
redis==4.6.0
python-dotenv==1.0.0
requests==2.31.0
# If PostgreSQL build tools are not available, you can use a pre-built binary:
# pip install --only-binary :all: psycopg2-binary==2.9.7
psycopg2-binary==2.9.7
structlog==23.1.0
PyJWT==2.8.0
# Dependency Injection
dependency-injector==4.48.1
# OpenTelemetry packages
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-exporter-prometheus==0.56b0
opentelemetry-instrumentation-flask==0.56b0
opentelemetry-instrumentation-requests==0.56b0
opentelemetry-instrumentation-redis==0.56b0
opentelemetry-instrumentation-psycopg2==0.56b0
# Type stubs
types-redis==4.6.0.7
# Optional dependencies
# python-thermoworks-cloud==1.0.0  # Uncomment if available from a private repository
